Relocate view models to Models.ViewModels folder
