+++ Support_Your_Locals/Models/ViewModels/ErrorViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace Support_Your_Locals.Models.ViewModels
+{
+    public class ErrorViewModel
+    {
+        public string RequestId { get; set; }
+
+        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+    }
+}
