+++ Support_Your_Locals/Models/ViewModels/UserLoginModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Support_Your_Locals.Models
+namespace Support_Your_Locals.Models.ViewModels
 {
     public class UserLoginModel
     {
