+++ Support_Your_Locals/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using Support_Your_Locals
-@using Support_Your_Locals.Models
+@using Support_Your_Locals.Models.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
